DL-113 fix the error of network and now trying to solve the error of sonar token

+++ Jenkinsfile
@@ -10,6 +10,7 @@ pipeline {
         SONAR_HOST_URL = "http://sonarqube:9000"
         SONAR_AUTH_TOKEN = credentials('sonar-token')
         TESTCONTAINERS_RYUK_DISABLED = "true"
+        SONAR_QUBE_CONFIG = 'sonarqube'
     }
 
     stages {
@@ -23,40 +24,32 @@ pipeline {
 
         stage('Build & Test') {
             steps {
-                echo 'Nettoyage de la configuration Testcontainers...'
-                sh 'rm -f $HOME/.testcontainers.properties'
-
-                echo 'Lancement de mvn clean verify...'
-                sh 'mvn clean verify'
+                script {
+                    if (fileExists("$HOME/.testcontainers.properties")) {
+                        echo "Nettoyage de l'ancienne configuration Testcontainers..."
+                        sh "rm -f $HOME/.testcontainers.properties"
+                    } else {
+                        echo "Aucun fichier de configuration Testcontainers à nettoyer."
+                    }
+                }
+                    echo 'Lancement de mvn clean verify...'
+                    sh 'mvn clean verify'
             }
         }
 
         stage('SonarQube Analysis') {
             steps {
-                script {
-                    if (currentBuild.result == null || currentBuild.result == 'SUCCESS') {
-                        echo 'Lancement de l\'analyse SonarQube...'
-                        withSonarQubeEnv('sonarqube') {
-                            sh 'mvn sonar:sonar'
-                        }
-                    } else {
-                        echo 'Tests échoués, on saute l\'analyse SonarQube.'
-                    }
+                echo "Lancement de l'analyse SonarQube..."
+                withSonarQubeEnv(SONAR_QUBE_CONFIG) {
+                sh 'mvn sonar:sonar -Dsonar.login=$SONAR_AUTH_TOKEN'
                 }
             }
         }
 
         stage('Quality Gate Check') {
             steps {
-                script {
-                    if (currentBuild.result == null || currentBuild.result == 'SUCCESS') {
-                        echo 'Vérification du statut du Quality Gate...'
-                        timeout(time: 5, unit: 'MINUTES') {
-                            waitForQualityGate abortPipeline: true
-                        }
-                    } else {
-                        echo 'Tests échoués, on saute le Quality Gate.'
-                    }
+                timeout(time: 1, unit: 'HOURS') {
+                    waitForQualityGate abortPipeline: true
                 }
             }
         }
@@ -65,6 +58,7 @@ pipeline {
     post {
         always {
             echo 'Fin du pipeline.'
+            cleanWs()
         }
     }
 }
